+++ src/main/java/domain/ladder/Ladder.java
@@ -1,5 +1,6 @@
-package domain;
+package domain.ladder;
 
+import domain.player.Players;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.function.BooleanSupplier;
