+++ src/main/java/domain/ladder/LadderHeight.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.ladder;
 
 public class LadderHeight {
