+++ src/main/java/domain/ladder/LadderRowPattern.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.ladder;
 
 import java.util.Collections;
 import java.util.List;
