+++ src/main/java/domain/player/Players.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.player;
 
 import java.util.List;
