+++ src/test/java/domain/LadderRowTest.java
@@ -3,6 +3,8 @@ package domain;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
+import domain.ladder.LadderRow;
+import domain.ladder.LadderRowPattern;
 import java.util.List;
 import java.util.function.BooleanSupplier;
 import org.junit.jupiter.api.DisplayName;
